+++ feature/condregister/maybe_taildrop.go
@@ -1,0 +1,8 @@
+// Copyright (c) Tailscale Inc & AUTHORS
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !ts_omit_taildrop
+
+package condregister
+
+import _ "tailscale.com/feature/taildrop"
