+++ feature/taildrop/doc.go
@@ -1,0 +1,5 @@
+// Copyright (c) Tailscale Inc & AUTHORS
+// SPDX-License-Identifier: BSD-3-Clause
+
+// Package taildrop registers the taildrop (file sending) feature.
+package taildrop
